+++ react_app/src/reducers/index.js
@@ -43,13 +43,6 @@ const courseInfoReducer = (courseInfo = {}, action) => {
     return courseInfo;
 }
 
-const contactFormReducer = (message = '', action) => {
-    if (action.type === 'SEND_CONTACT_FORM') {
-        return action.payload;
-    }
-    return message;
-}
-
 const allReducers = combineReducers({
     courseList: courseReducer,
     filters: filterReducer,
@@ -61,7 +54,7 @@ const allReducers = combineReducers({
     courseInfo: courseInfoReducer,
     countPage: countPage,
     currentPage: currentPage,
-    searchString: searchString,
+    searchString: searchString
 });
 
 const store = createStore(allReducers, applyMiddleware(thunk))
